requirements.txt
requests
msgpack-python